pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-code-mode-demo"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastmcp>=3.2.0",
  "mcp>=1.2.0",
  "anthropic>=0.40.0",
  "python-dotenv>=1.0.1",
  "flask>=3.0",
  "requests>=2.31",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[tool.setuptools.packages.find]
include = ["mlops_backend*", "servers*", "bench*", "tests*"]
exclude = ["codemodevenv*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
